Remove -fno-strictness workarounds
Thanks to https://github.com/input-output-hk/plutus/pull/3802 we can now
handle generated worker-wrapper splits! Hooray, one fewer incomprehensible error!

So we can go through and remove all the `-fno-strictness` and
`-fno-worker-wrapper` settings.

However, we now have the different problem that while the worker-wrapper
splits *work*, they tend to generate more code, which we don't benefit
from (since we're not avoiding (un)boxing).

So we get more code. I'm not really sure what to do about that. We could
skip this PR, and just say "if it crops up in user code that's okay, but
we'll still recommend turning it off to avoid code bloat". Or we could
just accept it.